+++ packages/ui/src/theme/index.ts
@@ -3,5 +3,5 @@ export * from "./types/editor";
 export * from "./types/theme";
 export * from "./utils/apply-theme";
 export * from "./utils/theme-preset-helper";
-export * from "./utils/theme-presets";
 export * from "./utils/theme-script";
+export * from "./presets";